refactor: use hooks for authorization
