+++ src/pages/LoginPage.jsx
@@ -12,22 +12,14 @@ import { useTheme, alpha } from '@mui/material/styles';
 import { NavbarComponent } from '../components/NavbarComponent';
 import { GlassBoxComponent } from '../components/GlassBoxComponent';
 import { useNavigate } from 'react-router-dom';
-import { useDispatch, useSelector } from 'react-redux';
-import { authAsyncAction } from '../store/authAsyncAction';
 import { Formik } from 'formik';
-import { useEffect } from 'react';
-import { setStatus } from '../store/slice';
+import { useAuth } from '../hooks/useAuth';
 
 export const LoginPage = () => {
   const theme = useTheme();
-  const dispatch = useDispatch();
   const nav = useNavigate();
-  const { status, user, error } = useSelector((state) => state.auth);
-  const isSuccess = status.user.login === 'SUCCESS';
 
-  const handleLogin = (values, { setSubmitting }) => {
-    dispatch(authAsyncAction.login(values)).finally(() => setSubmitting(false));
-  };
+  const { handleAuth, error, user } = useAuth('login');
 
   return (
     <Box
@@ -58,130 +50,116 @@ export const LoginPage = () => {
           <Typography variant="h5" fontWeight={700} color={theme.palette.text.primary} mb={2}>
             Login
           </Typography>
-          <Formik initialValues={{ email: '', password: '' }} onSubmit={handleLogin}>
-            {({ values, handleChange, handleBlur, handleSubmit, isSubmitting, resetForm }) => {
-              useEffect(() => {
-                if (isSuccess && user && user.id) {
-                  localStorage.setItem('userId', user.id);
-                  setTimeout(() => {
-                    nav('../dashboard');
-                    dispatch(
-                      setStatus({ type: 'user', status: { ...status.user, login: 'IDLE' } })
-                    );
-                    resetForm();
-                  }, 1000);
-                }
-              }, [isSuccess, nav, dispatch, resetForm, status.user, user]);
-              return (
-                <Box component="form" sx={{ width: '100%' }} onSubmit={handleSubmit}>
-                  <Box mb={2}>
-                    <FormControl
-                      fullWidth
-                      variant="outlined"
-                      sx={{
-                        mb: 1,
-                        borderRadius: 2,
-                        background: alpha(theme.palette.background.paper, 0.35),
-                      }}
+          <Formik initialValues={{ email: '', password: '' }} onSubmit={handleAuth}>
+            {({ values, handleChange, handleBlur, handleSubmit, isSubmitting, resetForm }) => (
+              <Box component="form" sx={{ width: '100%' }} onSubmit={handleSubmit}>
+                <Box mb={2}>
+                  <FormControl
+                    fullWidth
+                    variant="outlined"
+                    sx={{
+                      mb: 1,
+                      borderRadius: 2,
+                      background: alpha(theme.palette.background.paper, 0.35),
+                    }}
+                  >
+                    <InputLabel
+                      htmlFor="login-email"
+                      sx={{ color: theme.palette.text.secondary }}
                     >
-                      <InputLabel
-                        htmlFor="login-email"
-                        sx={{ color: theme.palette.text.secondary }}
-                      >
-                        Email
-                      </InputLabel>
-                      <OutlinedInput
-                        id="login-email"
-                        name="email"
-                        type="email"
-                        label="Email"
-                        value={values.email}
-                        onChange={handleChange}
-                        onBlur={handleBlur}
-                        sx={{
-                          color: theme.palette.text.primary,
-                          borderRadius: 2,
-                          '& .MuiOutlinedInput-notchedOutline': {
-                            borderColor: theme.palette.divider,
-                          },
-                          '&:hover .MuiOutlinedInput-notchedOutline': {
-                            borderColor: theme.palette.primary.main,
-                          },
-                          background: 'transparent',
-                        }}
-                      />
-                    </FormControl>
-                  </Box>
-                  <Box mb={3}>
-                    <FormControl
-                      fullWidth
-                      variant="outlined"
+                      Email
+                    </InputLabel>
+                    <OutlinedInput
+                      id="login-email"
+                      name="email"
+                      type="email"
+                      label="Email"
+                      value={values.email}
+                      onChange={handleChange}
+                      onBlur={handleBlur}
                       sx={{
-                        mb: 1,
+                        color: theme.palette.text.primary,
                         borderRadius: 2,
-                        background: alpha(theme.palette.background.paper, 0.35),
+                        '& .MuiOutlinedInput-notchedOutline': {
+                          borderColor: theme.palette.divider,
+                        },
+                        '&:hover .MuiOutlinedInput-notchedOutline': {
+                          borderColor: theme.palette.primary.main,
+                        },
+                        background: 'transparent',
                       }}
-                    >
-                      <InputLabel
-                        htmlFor="login-password"
-                        sx={{ color: theme.palette.text.secondary }}
-                      >
-                        Password
-                      </InputLabel>
-                      <OutlinedInput
-                        id="login-password"
-                        name="password"
-                        type="password"
-                        label="Password"
-                        value={values.password}
-                        onChange={handleChange}
-                        onBlur={handleBlur}
-                        sx={{
-                          color: theme.palette.text.primary,
-                          borderRadius: 2,
-                          '& .MuiOutlinedInput-notchedOutline': {
-                            borderColor: theme.palette.divider,
-                          },
-                          '&:hover .MuiOutlinedInput-notchedOutline': {
-                            borderColor: theme.palette.primary.main,
-                          },
-                          background: 'transparent',
-                        }}
-                      />
-                    </FormControl>
-                  </Box>
-                  <Box textAlign="center">
-                    <Button
-                      type="submit"
-                      variant="contained"
-                      color="primary"
-                      fullWidth
-                      disabled={isSubmitting}
-                      sx={{
-                        borderRadius: 2,
-                        fontWeight: 700,
-                        fontSize: 16,
-                        py: 1.5,
-                        boxShadow: theme.shadows[2],
-                        textTransform: 'none',
-                      }}
-                    >
-                      {isSubmitting ? 'Loading...' : 'Log In'}
-                    </Button>
-                    {error && (
-                      <Typography color="error" mt={2}>
-                        {error}
-                      </Typography>
-                    )}
-                    {user && (
-                      <Typography color="success.main" mt={2}>
-                        Login succeed!
-                      </Typography>
-                    )}
-                  </Box>
+                    />
+                  </FormControl>
                 </Box>
-              );
-            }}
+                <Box mb={3}>
+                  <FormControl
+                    fullWidth
+                    variant="outlined"
+                    sx={{
+                      mb: 1,
+                      borderRadius: 2,
+                      background: alpha(theme.palette.background.paper, 0.35),
+                    }}
+                  >
+                    <InputLabel
+                      htmlFor="login-password"
+                      sx={{ color: theme.palette.text.secondary }}
+                    >
+                      Password
+                    </InputLabel>
+                    <OutlinedInput
+                      id="login-password"
+                      name="password"
+                      type="password"
+                      label="Password"
+                      value={values.password}
+                      onChange={handleChange}
+                      onBlur={handleBlur}
+                      sx={{
+                        color: theme.palette.text.primary,
+                        borderRadius: 2,
+                        '& .MuiOutlinedInput-notchedOutline': {
+                          borderColor: theme.palette.divider,
+                        },
+                        '&:hover .MuiOutlinedInput-notchedOutline': {
+                          borderColor: theme.palette.primary.main,
+                        },
+                        background: 'transparent',
+                      }}
+                    />
+                  </FormControl>
+                </Box>
+                <Box textAlign="center">
+                  <Button
+                    type="submit"
+                    variant="contained"
+                    color="primary"
+                    fullWidth
+                    disabled={isSubmitting}
+                    sx={{
+                      borderRadius: 2,
+                      fontWeight: 700,
+                      fontSize: 16,
+                      py: 1.5,
+                      boxShadow: theme.shadows[2],
+                      textTransform: 'none',
+                    }}
+                  >
+                    {isSubmitting ? 'Loading...' : 'Log In'}
+                  </Button>
+                  {error && (
+                    <Typography color="error" mt={2}>
+                      {error}
+                    </Typography>
+                  )}
+                  {user && (
+                    <Typography color="success.main" mt={2}>
+                      Login succeed!
+                    </Typography>
+                  )}
+                </Box>
+              </Box>
+            )}
           </Formik>
           <Grid mt={2}>
             Didn't have an account? <Button onClick={() => nav('../signup')}>Sign Up</Button>
